refactor: split rust artifact plan modules
